requirements.txt
pillow
pytesseract
opencv-python
numpy